change: directory structure and imports a function used by each view
