+++ htcanalyze/log_summarizer/__init__.py
@@ -1,2 +1,4 @@
 
+
+from .summarized_condor_logs import SummarizedCondorLogs
 from .htcsummarizer import HTCSummarizer
